WIP Simplify server
