+++ server/lib/lib.rs
@@ -1,4 +1,6 @@
+pub mod handlers;
 pub mod opus;
+pub mod state;
 
 use candle_core::{Device, Tensor};
 use fish_speech_core::models::text2semantic::utils::encode::encode_tokens;
